Add "delete collection"

+++ app/src/main/kotlin/at/bitfire/davdroid/ui/account/DeleteCollectionDialog.kt
@@ -23,10 +23,13 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
+import androidx.compose.ui.window.DialogProperties
 import androidx.lifecycle.viewmodel.compose.viewModel
 import at.bitfire.davdroid.R
 import at.bitfire.davdroid.db.Collection
-import at.bitfire.davdroid.log.Logger
+import at.bitfire.davdroid.ui.ExceptionInfoDialog
+import okhttp3.HttpUrl.Companion.toHttpUrl
+import java.util.Optional
 
 @Composable
 fun DeleteCollectionDialog(
@@ -34,25 +37,39 @@ fun DeleteCollectionDialog(
     onDismiss: () -> Unit,
     model: AccountModel = viewModel()
 ) {
+    var started by remember { mutableStateOf(false) }
     val result by model.deleteCollectionResult.observeAsState()
-    result?.let { optionalResult ->
+
+    fun dismiss() {
+        // dismiss dialog, reset data so that it can be shown from start again
         onDismiss()
-        if (optionalResult.isPresent) {
-            // TODO exception
-            Logger.log.warning("ERROR")
-        }
+        started = false
+        model.deleteCollectionResult.value = null
     }
 
-    var deleting by remember { mutableStateOf(false) }
-    Dialog(onDismissRequest = onDismiss) {
+    if (result?.isEmpty == true) {
+        // finished without error
+        dismiss()
+        return
+    }
+
+    Dialog(
+        properties = DialogProperties(dismissOnClickOutside =
+            result != null ||       // finished with error message
+            !started                // not started
+        ),
+        onDismissRequest = ::dismiss
+    ) {
         Card {
             DeleteCollectionDialog_Content(
-                name = collection.title(),
+                collection = collection,
+                started = started,
+                result = result,
                 onDeleteCollection = {
-                    deleting = true
+                    started = true
                     model.deleteCollection(collection)
                 },
-                onDismiss = onDismiss
+                onCancel = ::dismiss
             )
         }
     }
@@ -60,10 +77,11 @@ fun DeleteCollectionDialog(
 
 @Composable
 fun DeleteCollectionDialog_Content(
-    name: String,
-    isDeleting: Boolean = false,
+    collection: Collection,
+    started: Boolean,
+    result: Optional<Exception>?,
     onDeleteCollection: () -> Unit = {},
-    onDismiss: () -> Unit = {}
+    onCancel: () -> Unit = {}
 ) {
     Column(Modifier.padding(16.dp)) {
         Text(
@@ -72,50 +90,75 @@ fun DeleteCollectionDialog_Content(
         )
 
         Text(
-            stringResource(R.string.delete_collection_confirm_warning, name),
+            stringResource(R.string.delete_collection_confirm_warning, collection.title()),
             style = MaterialTheme.typography.body1,
-            modifier = Modifier.padding(top = 8.dp)
+            modifier = Modifier.padding(vertical = 8.dp)
         )
 
-        if (isDeleting)
-            LinearProgressIndicator(
-                Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = 8.dp)
-            )
-        else {
-            var confirmed by remember { mutableStateOf(false) }
-            Row(Modifier.padding(vertical = 8.dp)) {
-                Checkbox(
-                    checked = confirmed,
-                    onCheckedChange = {
-                        confirmed = !confirmed
-                    }
-                )
-                Text(
-                    stringResource(R.string.delete_collection_data_shall_be_deleted),
-                    style = MaterialTheme.typography.body1,
-                    modifier = Modifier.clickable { confirmed = !confirmed }
+        when {
+            result != null -> {
+                val exception = result.get()
+                ExceptionInfoDialog(
+                    exception = exception,
+                    remoteResource = collection.url,
+                    onDismissRequest = onCancel
                 )
             }
 
-            Row(
-                horizontalArrangement = Arrangement.End,
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                TextButton(
-                    onClick = onDismiss
-                ) {
-                    Text(stringResource(android.R.string.cancel).uppercase())
-                }
-
-                TextButton(
-                    onClick = onDeleteCollection,
-                    enabled = confirmed
-                ) {
-                    Text(stringResource(R.string.delete_collection).uppercase())
-                }
+            started -> {
+                LinearProgressIndicator(
+                    color = MaterialTheme.colors.secondary,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(bottom = 8.dp)
+                )
             }
+
+            else ->
+                DeleteCollectionDialog_Confirmation(
+                    onDeleteCollection = onDeleteCollection,
+                    onCancel = onCancel
+                )
+        }
+
+    }
+}
+
+@Composable
+fun DeleteCollectionDialog_Confirmation(
+    onDeleteCollection: () -> Unit = {},
+    onCancel: () -> Unit = {}
+) {
+    var confirmed by remember { mutableStateOf(false) }
+    Row(Modifier.padding(vertical = 8.dp)) {
+        Checkbox(
+            checked = confirmed,
+            onCheckedChange = {
+                confirmed = !confirmed
+            }
+        )
+        Text(
+            stringResource(R.string.delete_collection_data_shall_be_deleted),
+            style = MaterialTheme.typography.body1,
+            modifier = Modifier.clickable { confirmed = !confirmed }
+        )
+    }
+
+    Row(
+        horizontalArrangement = Arrangement.End,
+        modifier = Modifier.fillMaxWidth()
+    ) {
+        TextButton(
+            onClick = onCancel
+        ) {
+            Text(stringResource(android.R.string.cancel).uppercase())
+        }
+
+        TextButton(
+            onClick = onDeleteCollection,
+            enabled = confirmed
+        ) {
+            Text(stringResource(R.string.delete_collection).uppercase())
         }
     }
 }
@@ -124,7 +167,12 @@ fun DeleteCollectionDialog_Content(
 @Preview
 fun DeleteCollectionDialog_Preview() {
     DeleteCollectionDialog_Content(
-        "Test Calendar"
+        collection = Collection(
+            type = Collection.TYPE_CALENDAR,
+            url = "https://example.com/calendar".toHttpUrl(),
+        ),
+        started = false,
+        result = null
     )
 }
 
@@ -132,7 +180,24 @@ fun DeleteCollectionDialog_Preview() {
 @Preview
 fun DeleteCollectionDialog_Preview_Deleting() {
     DeleteCollectionDialog_Content(
-        "Test Calendar",
-        isDeleting = true
+        collection = Collection(
+            type = Collection.TYPE_CALENDAR,
+            url = "https://example.com/calendar".toHttpUrl(),
+        ),
+        started = true,
+        result = null
+    )
+}
+
+@Composable
+@Preview
+fun DeleteCollectionDialog_Preview_Error() {
+    DeleteCollectionDialog_Content(
+        collection = Collection(
+            type = Collection.TYPE_CALENDAR,
+            url = "https://example.com/calendar".toHttpUrl(),
+        ),
+        started = true,
+        result = Optional.of(Exception("Test error"))
     )
 }
